Separate start function by entrypoints.
