+++ bids-validator/validators/bids/index.js
@@ -1,4 +1,3 @@
-import start from './start'
 import reset from './reset'
 import quickTest from './quickTest'
 import quickTestError from './quickTestError'
@@ -12,7 +11,6 @@ const subIDsesIDmismatchtest = subIDsesIDmismatchTest
 export {
   options,
   issues,
-  start,
   quickTestError,
   quickTest,
   fullTest,
@@ -23,7 +21,6 @@ export {
 export default {
   options,
   issues,
-  start,
   quickTestError,
   quickTest,
   fullTest,
